Refactor FinancialMovementRepository.

+++ src/main/java/com/lhama/lhamapersonalfinances/model/repositorys/FinancialMovementRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FinancialMovementRepository extends JpaRepository<FinancialMovement, Integer> {
+public interface FinancialMovementRepository extends JpaRepository<FinancialMovement, Long> {
 }
